+++ apps/demo/resources/views/examples/number-input/with-decimals.blade.php
@@ -1,0 +1,1 @@
+<x-ui.number-input name="price" :value="1.9" :step="0.1" :min="0" :decimals="2" :nullable="false" class="w-40" />